+++ src/main/java/com/accbdd/complicated_bees/block/MellariumEnergyCellBlock.java
@@ -1,13 +1,13 @@
 package com.accbdd.complicated_bees.block;
 
-import com.accbdd.complicated_bees.block.entity.mellarium.MellariumEnergyAcceptorBlockEntity;
+import com.accbdd.complicated_bees.block.entity.mellarium.MellariumEnergyCellBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class MellariumEnergyAcceptorBlock extends MellariumBlock {
+public class MellariumEnergyCellBlock extends MellariumBlock {
     @Override
     public BlockEntity newBlockEntity(BlockPos pPos, BlockState pState) {
-        return new MellariumEnergyAcceptorBlockEntity(pPos, pState);
+        return new MellariumEnergyCellBlockEntity(pPos, pState);
     }
 }