+++ changelog.md
@@ -3,7 +3,7 @@
 Add uses for power in the Mellarium! Like the Mellarium Skybox and Temporal Simulator!
 
 ## Added
-* Mellarium Energy Acceptor - #64
+* Mellarium Energy Cell - #64
 * Mellarium Skybox - #64
 * Mellarium Temporal Simulator - #64